DSE Basic UI (#48)
Adds a user interface for design space exploration configuration and
results. This synchronizes with the active / selected Run configuration,
but is on the block visualizer side-panel. Currently it's data display
only, a future PR will add a UI flow for adding / removing / modifying
search configurations and objective functions.

**Note: design space exploration is a work-in-progress feature.** It is
currently gated behind a feature flag since it is not complete and not
ready for general use.

The results display is a tree table listing all the objective values.
Future work might have customizable columns, but this is the quick and
dirty first iteration.

Specific changes:
- Create DseResult case class to store individual result points
- For CSV output of the DSE runner: create the empty CSV file at the
beginning (overwriting a prior run as needed), and write rows
incrementally
- Add TreeTableModels for search configurations, objective functions,
and results
- Aggregate results based on similarity - currently defined as same
objective function values and design (but not refinements)
- Add utility methods for cleaner display (toString) of search
configurations and values
- Raw values returned as part of the DSE config API (in addition to the
refinements) - values needed for cleaner display to the user
- Refactor the feature flag into a shared object across the UI and line
marker producer (gutter run icon)
- Refactor away DseReductionObjective - partly so that the type
parameter can be covariant, partly since it's a lot of specialized
infrastructure that wasn't conceptually coherent
- Serialization of search config and objective functions, so they can be
stored in the run config and persist across IDE restarts
- Serialization utils and unit tests - it's not perfect but should work
well enough
- Opportunistic cleanup
  - Add default values for restoring state