Show refdes in block diagram viewer (#54)
Show the refdes (for applicable blocks) in the block viewer, in addition
to the currently shown block name. If the block just contains a single
inner block with a refdes, shows that - this helps deal with the heavy
use of the adapter pattern (eg, DecouplingCapacitor contains a
Capacitor, that Capacitor is what has the refdes but it makes sense to
also show it on the wrapping DecouplingCapacitor).

This is done by using the mapper infrastructure provided in the
EdgirGraph -> ElkGraph pass, as an additional mapper that adds a Title
field to each ElkNode, which the conversion pass then reads and uses for
the title (if present). Mappers run for each block, checking for a
refdes (in the block, or an innermost only-child block), and if it
exists creates a title with it. Architecture-wise this is probably the
cleanest option, somewhat analogous to how DesignToolTipMap is
structured.

This also adds the mapper label infrastructure to ports, though that is
not currently used.

Works with both the current block visualizer panel and the PDFing pass.
I'd expect this to work with the recursive PDFing when that gets merged,
though this will probably add a few more lines to that merge conflict 😅

Other minor cleanup: remove the unneeded Project parameter from
DesignToolTipTextMap, use DesignPath.lastString instead of
reimplementing it.

Co-authored-by: TayKaiJun <tehohno@gmail.com>