Optional (no-longer-default) persist block cache (#61)
Make the block cache persist-across-IDE-restarts optional, configured by
a not-recommended checkbox in the project settings. The behavior is
really kind of flakey as of right now but useful in come cases for rapid
iteration.

Also renames to "Empty Block Cache" for better descriptiveness and rolls
the HDL submodule to include the cache behavior documentation.