DSE: Cleaning (#70)
DSE is a work-in-progress feature and not enabled by default.

Some incremental cleaning and polishing of the DSE feature.
- DSE plot: mouseover now displays data about a point, including
objective values and summary of all config points
- DSE plot: on selection also scroll to corresponding result row and
expand
- Ups the JVM heap, since large design spaces will cause it to blow past
memory limits.
- In the future this can be optimized by discarding the compiler, just
saving solved values.
- Use the new ParamToUnitsStringUtil for rendering values.
- Refactor ParamToUnitsStringUtil, to separate out the
numeric-value-units rendering from everything else, so it can be used as
a reusable function in other places
- In the future this (including units) might be better integrated into
infrastructure, but that's another refactor for another day
  - Eliminate trailing space when there is neither a SI prefix or units
- Unify DseObjectiveParameter, that now takes the Expr type as an
argument and returns a generic ExprValue. Specific unpacking is handled
at a higher level.