DSE: Support for plotting text values (#71)
DSE is a work-in-progress feature and not enabled by default.

Support plotting text in the plotting view, where text is mapped to
evenly spaced positions. Internal refactoring to support this:
- Updating data also allows specification of axes, as mapping of
positions to strings. If not specified, the dynamic numerical axes are
used. If specified, those points are displayed.
- Refactor the axis-selector internal model to map over all the results
instead of point-by-point, so that it can also return an axis definition
- Allow plotting non-numeric parameter values as text

Other changes:
- Aggregate points in the results tree-table by objective values only,
lumping different designs
- When the span (range of plotted values for any axis) is empty, default
to 1 for sizing purposes and to center the points
- When no axis is selected, default the position to zero (instead of
none) so at least a linear plot is shown