Hardware Proven Indicator (#72)
This is an experimental feature and is disabled by default. It can be
enabled through plugin settings.


![image](https://user-images.githubusercontent.com/1069863/220786441-5cb2f6b8-b4fb-4424-ab1c-99cbb60b49c6.png)

Adds a hardware-proven indicator next to libraries that includes the
count of this component that has been used in prior designs, and the
status (working / fixed / broken / untested). The data is pulled from a
CSV and design files within this repo, which contains a list of designs,
associated commit hash, and proven indicator of subpaths (so different
sub-blocks of the same top-level board can be marked as working /
broken).

Hovering over this indicator displays more details, showing the designs
(and how many times in the design) the block was used in.

The code in the future should flexibly pull from resources (in the JAR)
and user-configurable local directory.

Other changes
- Allow drag-and-drop of .edg files (containing schema.Design proto -
this is a recent change in the HDL repo) to visualize the saved file.
Currently does not display solved values or top-level refinements, that
needs a separate traversal of the design tree that isn't needed yet.
This is primarily a developer / internal feature and isn't meant to be
discoverable.
- Refactor library tree nodes to be inner classes of an outer class that
makes the library object and project available to all internal nodes,
instead of plumbing things through arguments.

TODO:
- [x] add to design view
- [x] display notes

Follow-on PR:
- Sequence the data from git and show if the class has been modified
since the last proven