Reduce startup time indexing KiCad footprints (#76)
Previously, this took about ~6000ms, this change gets it down to ~800ms.
With no footprints directory specified, the time to construct the model
and TreeTable is about 3ms. Probably the primary mechanism is the
simplified indexing logic plus using direct File from parent File
creation (as opposed to presumably a bunch of long string ops)

It looks like for whatever reason, on initialization the model is
traversed to the leaf level, even though the node children is
lazy-initialized. A future change (micro-optimization) might be to
create a TreeTable that can make use of lazily initialized children.
Thought may be needed on how this interacts with filtering, which
depends on the leaf nodes being available - maybe lazy index until
filtering is asked for, then display a spinny icon as filtering indexes
everything?

Also makes the footprint UI consistent with the rest, by putting the
list view on the left (instead of prior on the right). Also uses
JBScrollPane for UI look and feel consistency.