Experimental: insertion live templates for blocks (#111)
Adds insertion live templates, for block insertions only, as an
alternative to caret-based editing. Currently gated behind an
experimental setting in the plugin configurations.

Future PRs may clean up this feature and expand this to other constructs
like connections.

A future PR will eliminate the option for multiple edit modes, but for
now this is experimental and opt-in.

Also expands the library blocks sub-tree by default.